+++ configs/run.yaml
@@ -4,8 +4,8 @@
 # order of defaults determines the order in which configs override each other
 defaults:
   - _self_
-  - dataset: graph/cocitation_citeseer
-  - model: graph/gcn
+  - dataset: hypergraph/cocitation_cora
+  - model: hypergraph/unignn2
   - transforms: ${get_default_transform:${dataset},${model}} #no_transform
   - optimizer: default
   - loss: default
